[wpiutil] Refactor SpanMatcher and TestPrinters from ntcore
