+++ ntcore/src/test/native/cpp/net/ServerImplTest.cpp
@@ -10,10 +10,10 @@
 #include <vector>
 
 #include <gtest/gtest.h>
+#include <wpi/SpanMatcher.h>
 
 #include "../MockLogger.h"
 #include "../PubSubOptionsMatcher.h"
-#include "../SpanMatcher.h"
 #include "../TestPrinters.h"
 #include "../ValueMatcher.h"
 #include "Handle.h"
