+++ ntcore/src/test/native/cpp/net/WireEncoderTest.cpp
@@ -10,8 +10,8 @@
 #include <gtest/gtest.h>
 #include <wpi/json.h>
 #include <wpi/raw_ostream.h>
+#include <wpi/SpanMatcher.h>
 
-#include "../SpanMatcher.h"
 #include "../TestPrinters.h"
 #include "Handle.h"
 #include "PubSubOptions.h"
